fix: resolve KeyboardAccessibilityProvider ReferenceError that crashes app at runtime
